backend/requirements.txt
fastapi
uvicorn
groq
docker
python-dotenv
pydantic
langgraph